Cast font size to int for tc-lib-pdf-font lowest compatibility
The lowest tc-lib-pdf-font release (2.6.4) types Stack::insert() $size as ?int,
while newer releases accept float. Cast to int so the prefer-lowest build works
across the whole supported range.

+++ src/Pdf/Engine/TcLibPdfEngine.php
@@ -79,7 +79,7 @@ class TcLibPdfEngine extends AbstractPdfEngine
             $tcpdf->pon,
             (string)($font['family'] ?? 'helvetica'),
             (string)($font['style'] ?? ''),
-            (float)($font['size'] ?? 10),
+            (int)($font['size'] ?? 10),
         );
 
         $tcpdf->addPage($this->_pageOptions());
